server/requirements.txt
fastapi
uvicorn==0.20.0
sqlmodel
tenacity==8.2.3
yt-dlp==2023.10.13
moviepy==1.0.3
mediapipe==0.10.8
opencv-python
pydantic-settings
python-multipart==0.0.6
httpx
apscheduler
structlog==23.2.0
faster-whisper==0.10.0
numpy==1.26.2
scipy==1.11.4

google-generativeai
edge-tts
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
